# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Room 911",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://servicios.asstiseguridadsocial.com"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/auth/entry_room_911": {
            "post": {
                "tags": [
                    "Income record"
                ],
                "summary": "Registrar intentos de ingresos a Room-911",
                "operationId": "78ece1f5f2ddaada000d02202a6971f3",
                "parameters": [
                    {
                        "name": "identification",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Empleado autorizado"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/income_record/{id}": {
            "get": {
                "tags": [
                    "Income record"
                ],
                "summary": "Consultar registro de ingresos de un empleado",
                "operationId": "1e9f5e6ba7ba74b19fb0dc259d2f9dbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "income_record": []
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/v1/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar roles",
                "operationId": "b69983d611cdc667d934cd869a8cbfbd",
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "roles": []
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear rol",
                "operationId": "5c301fc0b488a85f72f0855e26da17de",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Rol creado con éxito"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/roles/{id}": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear rol",
                "operationId": "b6f20a9582c122f092a72e3d5b379e8e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": [
                                        "Rol actualizado con éxto"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "Mensaje de error"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}